+++ pkg/rpc/cdnsystem/cdnsystem.proto
@@ -47,8 +47,8 @@ message PieceSeed{
 
 // CDN System RPC Service
 service Seeder{
-  // generate seeds and return to scheduler
+  // Generate seeds and return to scheduler
   rpc ObtainSeeds(SeedRequest)returns(stream PieceSeed);
-  // get piece tasks from cdn
+  // Get piece tasks from cdn
   rpc GetPieceTasks(base.PieceTaskRequest)returns(base.PiecePacket);
 }
