+++ pkg/rpc/cdnsystem/server/server.go
@@ -33,9 +33,11 @@ import (
 	"google.golang.org/grpc/peer"
 )
 
-// SeederServer see cdnsystem.SeederServer
+// SeederServer  refer to cdnsystem.SeederServer
 type SeederServer interface {
+	// Generate seeds and return to scheduler
 	ObtainSeeds(context.Context, *cdnsystem.SeedRequest, chan<- *cdnsystem.PieceSeed) error
+	// Get piece tasks from cdn
 	GetPieceTasks(context.Context, *base.PieceTaskRequest) (*base.PiecePacket, error)
 }
