--- pkg/rpc/server_options.go
@@ -1,51 +1,0 @@
-/*
- *     Copyright 2020 The Dragonfly Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package rpc
-
-import (
-	"time"
-
-	logger "d7y.io/dragonfly/v2/internal/dflog"
-
-	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
-	grpc_zap "github.com/grpc-ecosystem/go-grpc-middleware/logging/zap"
-	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/keepalive"
-)
-
-var DefaultServerOptions = []grpc.ServerOption{
-	grpc.ConnectionTimeout(10 * time.Second),
-	grpc.InitialConnWindowSize(8 * 1024 * 1024),
-	grpc.KeepaliveEnforcementPolicy(keepalive.EnforcementPolicy{
-		MinTime: 1 * time.Minute,
-	}),
-	grpc.KeepaliveParams(keepalive.ServerParameters{
-		MaxConnectionIdle: 5 * time.Minute,
-	}),
-	grpc.MaxConcurrentStreams(100),
-	grpc.StreamInterceptor(grpc_middleware.ChainStreamServer(
-		streamServerInterceptor,
-		grpc_prometheus.StreamServerInterceptor,
-		grpc_zap.StreamServerInterceptor(logger.GrpcLogger.Desugar()),
-	)),
-	grpc.UnaryInterceptor(grpc_middleware.ChainUnaryServer(
-		unaryServerInterceptor,
-		grpc_prometheus.UnaryServerInterceptor,
-		grpc_zap.UnaryServerInterceptor(logger.GrpcLogger.Desugar()),
-	)),
-}
